refactor: Repository를 테스트 가능한 구조로 변경한다.
