+++ app/src/main/java/woowacourse/shopping/data/DefaultProductRepository.kt
@@ -1,8 +1,9 @@
 package woowacourse.shopping.data
 
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.ProductRepository
 
-class ProductRepository {
+class DefaultProductRepository : ProductRepository {
 
     private val products: List<Product> = listOf(
         Product(
@@ -22,7 +23,7 @@ class ProductRepository {
         ),
     )
 
-    fun getAllProducts(): List<Product> {
+    override fun getAllProducts(): List<Product> {
         return products
     }
 }
