+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -3,9 +3,9 @@ package woowacourse.shopping.ui
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.CartRepository
-import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.model.CartRepository
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.ProductRepository
 
 class MainViewModel(
     private val productRepository: ProductRepository,
